move the api files into an underscored directory to support the api/gem
