--- atomic-red-team/atomic_execution_template.html.erb
@@ -1,1 +1,0 @@
-TBD
