+++ atomic_red_team/atomic_execution_template.html.erb
@@ -1,0 +1,1 @@
+TBD
