+++ atomic_red_team/atomic_test_template.yaml
@@ -1,0 +1,26 @@
+---
+attack_technique: TODO
+display_name: TODO
+
+atomic_tests:
+- name: TODO
+  description: |
+    TODO
+
+  supported_platforms:
+    - windows
+    - macos
+    - centos
+    - ubuntu
+    - linux
+
+  input_arguments:
+    output_file:
+      description: TODO
+      type: todo
+      default: TODO
+
+  executor:
+    name: command_prompt
+    command: |
+      TODO
